# openapi.json
{
  "components": {
    "schemas": {
...
      "ChapterDto": {
        "properties": {
          "number": {
            "format": "double",
            "type": "number"
          },
          "published_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "read": {
            "description": "Whether the requesting user has read this chapter (number ≤ their progress).\n`None` for anonymous requests; `Some(bool)` when a valid `Bearer` token is present\n(frontend §9.2 auth-scoped read-state).",
            "type": [
              "boolean",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "url": {
            "description": "Resolved absolute URL to open the chapter page on the provider.",
            "type": "string"
          }
        },
        "required": [
          "number",
          "url"
        ],
        "type": "object"
      },
...
      "ContentType": {
        "description": "The medium/origin classification of a work.",
        "enum": [
          "manga",
          "manhwa",
          "manhua",
          "webtoon",
          "unknown"
        ],
        "type": "string"
      },
      "ContinueItem": {
        "properties": {
          "cover_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_read_number": {
            "format": "double",
            "type": "number"
          },
          "next_number": {
            "description": "The lowest unread chapter number above the user's progress, if any.",
            "format": "double",
            "type": [
              "number",
              "null"
            ]
          },
          "series_id": {
            "$ref": "#/components/schemas/SeriesId"
          },
          "series_title": {
            "type": "string"
          },
          "unread": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "series_id",
          "series_title",
          "last_read_number",
          "unread"
        ],
        "type": "object"
      },
...
      "MeStats": {
        "description": "Lifetime reading stats for the Home/Profile headline (frontend §9.3 `GET /v1/me/stats`).\n`chapters_read` is the sum of whole chapters below each series' last-read marker — an\nhonest proxy over stored progress (there is no per-chapter read-event log, so a daily\n\"streak\" is intentionally omitted rather than fabricated).",
        "properties": {
          "chapters_read": {
            "format": "int64",
            "type": "integer"
          },
          "completed": {
            "format": "int64",
            "type": "integer"
          },
          "reading": {
            "format": "int64",
            "type": "integer"
          },
          "tracking": {
            "format": "int64",
            "type": "integer"
          },
          "unread": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "tracking",
          "reading",
          "completed",
          "chapters_read",
          "unread"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "description": "RFC 9457 `application/problem+json` error body shape produced by [`ApiError`]. Declared\npurely for `OpenAPI` documentation — [`ApiError::into_response`] builds the JSON by hand so\nruntime callers never construct this type.",
        "example": {
          "detail": "resource not found",
          "status": 404,
          "title": "not_found",
          "type": "about:blank#not_found"
        },
        "properties": {
          "detail": {
            "type": "string"
          },
          "status": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "type": "object"
      },
      "ProfileDto": {
        "properties": {
          "email": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "username",
          "role"
        ],
        "type": "object"
      },
      "ProfileUpdate": {
        "properties": {
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
      "SeriesId": {
        "description": "Identifies a canonical [`crate::entities::Series`].",
        "format": "uuid",
        "type": "string"
      },
      "SeriesSourceId": {
        "description": "Identifies a [`crate::entities::SeriesSource`].",
        "format": "uuid",
        "type": "string"
      },
      "SeriesStatus": {
        "description": "Publication status of a canonical series.",
        "enum": [
          "ongoing",
          "completed",
          "hiatus",
          "cancelled",
          "unknown"
        ],
        "type": "string"
      },
      "SeriesSummary": {
        "properties": {
          "content_type": {
            "$ref": "#/components/schemas/ContentType"
          },
          "cover_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "$ref": "#/components/schemas/SeriesId"
          },
          "source_count": {
            "format": "int64",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/SeriesStatus"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "content_type",
          "status",
          "source_count"
        ],
        "type": "object"
      },
      "SessionDto": {
        "properties": {
          "created_at": {
            "type": "string"
          },
          "expires_at": {
            "type": "string"
          },
          "family_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "family_id",
          "created_at",
          "expires_at"
        ],
        "type": "object"
      },
...
      "WatchStatus": {
        "description": "A user's tracking status for a series.",
        "enum": [
          "reading",
          "planned",
          "completed",
          "dropped",
          "paused"
        ],
        "type": "string"
      },
      "WatchlistItem": {
        "properties": {
          "added_at": {
            "type": "string"
          },
          "cover_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_read_number": {
            "description": "The user's last-read chapter number, if any.",
            "format": "double",
            "type": [
              "number",
              "null"
            ]
          },
          "notify": {
            "type": "boolean"
          },
          "series_id": {
            "$ref": "#/components/schemas/SeriesId"
          },
          "series_title": {
            "description": "Embedded series title so the Watchlist board renders without a per-card detail\nfetch (frontend §9.3, kills the N+1).",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/WatchStatus"
          },
          "sync_excluded": {
            "description": "Whether this series is opted out of external sync (design v2 §A.5).",
            "type": "boolean"
          },
          "unread": {
            "description": "Unread chapters above the user's progress.",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "series_id",
          "status",
          "notify",
          "added_at",
          "series_title",
          "unread",
          "sync_excluded"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "TankoVault contributors"
    },
    "description": "Axum HTTP edge: public read/write + admin, auth, resolved links, SSE.",
    "license": {
      "name": ""
    },
    "title": "tankovault-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/me/continue": {
      "get": {
        "description": "Continue-reading cards for Home / the Series CTA (frontend §9.3): tracked, in-progress\nseries that have unread chapters, freshest activity first.",
        "operationId": "continue_reading",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ContinueItem"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Up to 24 continue-reading cards"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "authentication required"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "Get continue-reading cards",
        "tags": [
          "me-dashboard"
        ]
      }
    },
...
    "/v1/me/notification-prefs": {
      "get": {
        "description": "The caller's notification preferences JSON (frontend §9.4). `{}` means \"product defaults\".",
        "operationId": "notification_prefs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Product-defined free-form preferences JSON"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "authentication required"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "Get notification preferences",
        "tags": [
          "me-notifications"
        ]
      },
      "put": {
        "description": "Replace the caller's notification preferences (frontend §9.4). The body is stored verbatim\nas an open JSON document.",
        "operationId": "put_notification_prefs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "The stored preferences, echoed back"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "authentication required"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "Replace notification preferences",
        "tags": [
          "me-notifications"
        ]
      }
    },
...
    "/v1/me/profile": {
      "patch": {
        "description": "Update the caller's username and/or email (frontend §9.4). A duplicate email/username\nsurfaces as `409 Conflict`.",
        "operationId": "patch_profile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            },
            "description": "Updated profile"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "authentication required"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Email or username already taken"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "Update the profile",
        "tags": [
          "me-account"
        ]
      }
    },
...
    "/v1/me/recommendations": {
      "get": {
        "description": "*Stub*: unwatched series sharing tags with the user's list (frontend §9.3). Falls back to\nthe most-recent catalog when the user has no tagged watchlist yet, so the shelf is never\nempty for signed-in users.",
        "operationId": "recommendations",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SeriesSummary"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Up to 12 recommended series"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "authentication required"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "Get \"because you read\" recommendations",
        "tags": [
          "me-dashboard"
        ]
      }
    },
    "/v1/me/sessions": {
      "get": {
        "description": "The caller's active login sessions (frontend §9.4).",
        "operationId": "sessions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Active sessions"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "authentication required"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "List active sessions",
        "tags": [
          "me-account"
        ]
      }
    },
    "/v1/me/sessions/{id}": {
      "delete": {
        "description": "Revoke one of the caller's own sessions (frontend §9.4). Scoped to ownership; a\nforeign/unknown id yields `404`.",
        "operationId": "delete_session",
        "parameters": [
          {
            "description": "Session id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "revoked": 1
                },
                "schema": {}
              }
            },
            "description": "Revoked"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "No such session for this caller"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "Revoke a session",
        "tags": [
          "me-account"
        ]
      }
    },
    "/v1/me/stats": {
      "get": {
        "description": "*Stub*: lifetime tracking stats for the Home / Profile headline (frontend §9.3). See\n[`tankovault_db::repo::tracking::MeStats`] for the honest definition of `chapters_read` and\nwhy no \"streak\" is returned.",
        "operationId": "stats",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeStats"
                }
              }
            },
            "description": "Lifetime stats"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "authentication required"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "Get lifetime tracking stats",
        "tags": [
          "me-dashboard"
        ]
      }
    },
...
    "/v1/me/watchlist": {
      "get": {
        "description": "The user's watchlist with embedded title/cover/progress (frontend §9.3). Extra fields are\nadditive; older clients ignore them.",
        "operationId": "watchlist",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WatchlistItem"
                  },
                  "type": "array"
                }
              }
            },
            "description": "The caller's watchlist"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "authentication required"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "Get the watchlist",
        "tags": [
          "me-watchlist"
        ]
      }
    },
...
    "/v1/series/{id}/chapters": {
      "get": {
        "description": "Chapter list, newest first. When a valid access token is supplied the per-chapter `read`\nflag is populated from the user's progress (frontend §9.2 auth-scoped read-state);\nanonymous callers get the same list without read-state.",
        "operationId": "chapters",
        "parameters": [
          {
            "description": "Series id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SeriesId"
            }
          },
          {
            "description": "Which source to read chapters from. Defaults to the first source of the series.",
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SeriesSourceId"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChapterDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Chapters, newest first"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Series (or its source) not found"
          }
        },
        "security": [
          {},
          {
            "bearer_auth": []
          }
        ],
        "summary": "List a source's chapters",
        "tags": [
          "series"
        ]
      }
    },
...
  },
...
}